+++ lib_xua/src/core/endpoint0/descriptor_defs.h
@@ -1,7 +1,12 @@
-// Copyright (c) 2015-2018, XMOS Ltd, All rights reserved
+// Copyright (c) 2015-2019, XMOS Ltd, All rights reserved
 
 #ifndef __DESCRIPTOR_DEFS_H__
 #define __DESCRIPTOR_DEFS_H__
+
+/*
+    Include xua.h to pick up the #defines of NUM_USB_CHAN_IN and NUM_USB_CHAN_OUT.
+ */
+#include "xua.h"
 
 #if (NUM_USB_CHAN_IN > 0) && (NUM_USB_CHAN_OUT > 0)
 #define AUDIO_INTERFACE_COUNT 3
@@ -54,10 +59,14 @@ enum USBInterfaceNumber
     INTERFACE_NUMBER_IAP_EA_NATIVE_TRANS,
 #endif
 #endif
-#if defined(HID_CONTROLS) && (HID_CONTROLS != 0)
+#if( 0 < HID_CONTROLS )
     INTERFACE_NUMBER_HID,
 #endif
     INTERFACE_COUNT          /* End marker */
 };
+
+#if( 0 < HID_CONTROLS )
+#define ENDPOINT_INT_INTERVAL_IN_HID 0x08
+#endif
 
 #endif