+++ CHANGELOG.rst
@@ -1,24 +1,81 @@
 lib_xua Change Log
 ==================
 
+1.1.1
+-----
+
+  * RESOLVED: Zero length input packets generated before enumeration causing I2S
+    timing pushout at startup
+  * CHANGED: Pin Python package versions
+  * REMOVED: not necessary cpanfile
+
+1.1.0
+-----
+
+  * ADDED:     Ability to read or modify serial number string
+
+1.0.1
+-----
+
+  * FIXED:     Wrong size of vendor and product strings
+
+1.0.0
+-----
+
+  * ADDED:     UAC1 HID support with simulated Voice Command detection reported
+    every 10 seconds
+  * ADDED:     Support for USB HID Set Idle request
+  * ADDED:     Pre-processor symbols to enable single-threaded, dual-PDM
+    microphone operation
+  * FIXED:     Descriptors for XUA_ADAPTIVE incorrectly defined for IN endpoint
+  * ADDED:     Guards to user_hid.h and xua_hid.h
+  * ADDED:     UAC1 HID support for AC Stop (End Call), Volume Increment and
+    Volume Decrement
+  * CHANGE:    UAC1 HID to report function keys f21 through f24 as specified by
+    customer
+  * CHANGE:    HID interface for user to set and clear events from global
+    variable to function
+  * CHANGE     HID report descriptor to use generic events instead of GPI
+    events, to report Key-phrase detection as AC Search, and to report end-call
+    detection as AC Stop
+  * ADDED:     Ability to read or modify vendor and product IDs and strings
+  * ADDED:     Ability to read or modify bcdDevice
+  * ADDED:     Override USB descriptor with sampling frequency and
+    bit-resolution set at boot time.
+  * ADDED:     Global pointer to allow external access to masterClockFreq
+
+0.2.1
+-----
+
+  * HOTFIX: Fix descriptors for XUA_ADAPTIVE
+
 0.2.0
 -----
 
   * ADDED:     Initial library documentation
   * ADDED:     Application note AN00247: Using lib_xua with lib_spdif (transmit)
+  * ADDED:     Separate callbacks for input/output audio stream start/stop
   * CHANGE:    I2S hardware resources no longer used globally and must be passed
     to XUA_AudioHub()
   * CHANGE:    XUA_AudioHub() no longer pars S/PDIF transmitter task
   * CHANGE:    Moved to lib_spdif (from module_spdif_tx & module_spdif_rx)
   * CHANGE:    Define NUM_PDM_MICS renamed to XUA_NUM_PDM_MICS
   * CHANGE:    Define NO_USB renamed to XUA_USB_EN
+  * CHANGE:    Build files updated to support new "xcommon" behaviour in xwaf.
   * RESOLVED:  wChannelConfig in UAC1 descriptor set according to output channel
     count
   * RESOLVED:  Indexing of ADAT channel strings (#18059)
+  * RESOLVED:  Rebooting device fails when PLL config "not reset" bit is set
 
   * Changes to dependencies:
 
-    - lib_spdif: Added dependency 3.0.0
+    - lib_dsp: Added dependency 5.0.0
+
+    - lib_mic_array: Added dependency 4.0.0
+
+    - lib_spdif: Added dependency 3.1.0
+
+    - lib_xassert: Added dependency 3.0.1
 
 0.1.2
 -----
@@ -92,7 +149,7 @@ Legacy release history
 
 7.3.0
 -----
-    - CHANGE:     Example OSX DFU host app updated to now take PID as runtime 
+    - CHANGE:     Example OSX DFU host app updated to now take PID as runtime
       argument. This enabled multiple XMOS devices to be attached to the host
       during DFU process